+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Browse | The request doesn't send when user clicks on highlighted in bold contributor name. Fixes UIIN-2140.
 * Incorrect format of the missing match in the call number browse. Refs UIIN-2141.
 * Increase limit for fetching child/parents relations. Fixes UIIN-2150.
+* The browse query changed when user returns from search to browse contributors pane. UIIN-2125.
 
 ## [9.1.0](https://github.com/folio-org/ui-inventory/tree/v9.1.0) (2022-06-28)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.0.0...v9.1.0)